% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrosselGraficos.R
\name{carrosselGraficos}
\alias{carrosselGraficos}
\title{Printa todos os gráficos de uma lista recursivamente!}
\usage{
carrosselGraficos(lista_graficos, iterativo = FALSE, nome_var = "")
}
\arguments{
\item{lista_graficos}{Uma lista com gráficos.}

\item{iterativo}{Para a cada gráfico printado e espera o usuário apertar "Enter".}

\item{nome_var}{Um nome que será usado no nome da variável no print no prompt.}
}
\description{
Printa todos os gráficos de uma lista recursivamente!
}
