% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isVecEqual.R
\name{isVecEqual}
\alias{isVecEqual}
\title{Compara se dois vetores possuem os mesmos elementos sem se importar com a ordem}
\usage{
isVecEqual(v1, v2)
}
\arguments{
\item{v1}{Vetor 1.}

\item{v2}{Vetor 2.}
}
\value{
TRUE ou FALSE
}
\description{
Compara se dois vetores possuem os mesmos elementos sem se importar com a ordem
}
