% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomeArqs.padroniza.R
\name{nomeArqs.padroniza}
\alias{nomeArqs.padroniza}
\title{Padroniza a numeração dos nomes dos arquivos.}
\usage{
nomeArqs.padroniza(caminho = "R", recursivo = TRUE)
}
\arguments{
\item{caminho}{O caminho (pasta) cujo arquivos se deseja padronizar o nome.}

\item{recursivo}{Se os nomes devem ser padronizados recursivamente.}
}
\description{
Padroniza a numeração dos nomes dos arquivos.
}
