% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomeArqs.somaUm.R
\name{nomeArqs.somaUm}
\alias{nomeArqs.somaUm}
\title{Soma 1 na numeração do nome dos arquivos a partir de um determinado número.}
\usage{
nomeArqs.somaUm(numero, caminho)
}
\arguments{
\item{numero}{A partir deste número (incluso), a numeração dos arquivos será adicionada de 1.}

\item{caminho}{O caminho (pasta) cujo arquivos se deseja adicionar 1 no nome.}
}
\description{
Soma 1 na numeração do nome dos arquivos a partir de um determinado número.
}
