% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvarTabelas.R
\name{salvarTabelas}
\alias{salvarTabelas}
\title{Salve as tabelas da sua lista}
\usage{
salvarTabelas(tabelas_list, save_path)
}
\arguments{
\item{tabelas_list}{Uma lista com tabelas (salva recursivo).}

\item{save_path}{Pasta para salvar os gráficos.}
}
\description{
Salve as tabelas da sua lista
}
