% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecionar.R
\name{selecionar}
\alias{selecionar}
\title{Selecionar: seleciona e altera o nome!}
\usage{
selecionar(dados, ...)
}
\arguments{
\item{dados}{Um data frame}

\item{...}{Argumentos do tipo expr(\emph{Nome}) = '\emph{nome}', onde \emph{Nome} é o nome de display da coluna, e '\emph{nome}' é o nome da coluna que será renomeada.}
}
\value{
Retorna o data frame com as colunas selecionadas e os nomes atualizados
}
\description{
Seleciona e altera o nome das colunas de um data frame utilizando uma \emph{expr}.
}
\examples{
df <- datasets::iris
COL_SEPAL_LENGTH <- rlang::expr(`Sepal Length`)
selecionar(df, COL_SEPAL_LENGTH = 'Sepal.Length')
}
