% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvarGraficos.R
\name{salvarGraficos}
\alias{salvarGraficos}
\title{Salve os gr치ficos da sua lista}
\usage{
salvarGraficos(lista_graficos, save_path)
}
\arguments{
\item{lista_graficos}{Uma lista com gr치ficos (salva recursivo).}

\item{save_path}{Pasta para salvar os gr치ficos.}
}
\description{
Salve os gr치ficos da sua lista
}
