% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvarTabelas.R
\name{salvarTabelas}
\alias{salvarTabelas}
\title{Salve as tabelas da sua lista}
\usage{
salvarTabelas(
  lista_tabelas,
  formato = "csv",
  path = "output/tabelas",
  progresso = ""
)
}
\arguments{
\item{formato}{'csv' ou 'xlsx'.}

\item{progresso}{Parâmetro usado pela recursão. Não deve ser utilizado}

\item{tabelas_list}{Uma lista com tabelas (salva recursivo).}

\item{save_path}{Pasta para salvar as tabelas.}
}
\description{
Salve as tabelas da sua lista
}
